+++ src/mefit/mefit.csproj
@@ -98,7 +98,7 @@
     <Compile Include="Firmware\SOCROM\SOCSigs.cs" />
     <Compile Include="Forms\Prompts\Prompts.cs" />
     <Compile Include="Forms\Types\Enums.cs" />
-    <Compile Include="MemoryTracker.cs" />
+    <Compile Include="Common\MemoryTracker.cs" />
     <Compile Include="UI\AccentColorHelper.cs" />
     <Compile Include="UI\BlurHelper.cs" />
     <Compile Include="UI\METForm.cs">
